+++ workshop1/eca1_jax_copy.py
@@ -36,7 +36,7 @@ def main(
     print("initialising state...")
     match init:
         case "middle":
-            state = jnp.zeros(width, dtype=int)
+            state = jnp.zeros(width, dtype=jnp.uint8)
             state = state.at[width//2].set(1)
         case "random":
             # jnp.random.seed(seed) # wrong: it's jax.random not
@@ -48,7 +48,7 @@ def main(
                 minval=0,
                 maxval=2, # not included
                 shape=(width,),
-                dtype=int,
+                dtype=jnp.uint8,
             )
     print("initial state:", state)
 
@@ -82,20 +82,21 @@ def main(
         
 def simulate(
     rule: int,
-    init_state: jax.Array,    # int[width]
+    init_state: jax.Array,    # uint8[width]
     height: int,
-) -> jax.Array:               # int[height, width]
+) -> jax.Array:               # uint8[height, width]
     # parse rule
     rule_uint8 = jnp.uint8(rule)
     rule_bits = jnp.unpackbits(rule_uint8, bitorder='little')
     rule_table = rule_bits.reshape(2,2,2)
 
     # parse initial state
+    init_state = init_state.astype(jnp.uint8)
     (width,) = init_state.shape
 
     # accumulate output into this array
     # extra width is to implement wraparound with slicing
-    history = jnp.zeros((height, width+2), dtype=int)
+    history = jnp.zeros((height, width+2), dtype=jnp.uint8)
 
     # first row
     history = history.at[0, 1:-1].set(init_state)
